Hotfix: Copy ENERGY to PI column in ero_calevents
The eSASS requires a PI column in unit eV. This is different from the
typical handling of the PI column in SIXTE, which is (e.g., for the
Athena WFI) on the channel grid and in unit ADU. As a hack the energy
column can be copied into PI but note that this bypasses a proper
PHA2PI correction!
This should be treated as pure hotfix to get approximate spectra and
should be treated with care!

+++ tools/ero_calevents/ero_calevents.h
@@ -71,7 +71,7 @@ typedef struct {
   long frame;
 
   long pha; /* [adu] */
-  long pi; /* currently no pha2pi correction --> contains same as pha */
+  float pi; /* currently no pha2pi correction! */
   float energy; /* [eV] */
 
   int rawx, rawy;
